move theme from User -> Colleague
